+++ src/linux/ConnectionImpl.h
@@ -7,7 +7,7 @@
 
 class ConnectionImpl{
 public:
-    ConnectionImpl(const AMQP::Address& address);
+    ConnectionImpl(const AMQP::Address& address, int connectTimeoutSec);
     virtual ~ConnectionImpl();
     void connect();
     AMQP::Channel* channel();
@@ -28,4 +28,5 @@ private:
     std::unique_ptr<AMQP::TcpChannel> rcChannel;
     std::thread thread;
     volatile bool stop;
+    int connectTimeoutSec;
 };