+++ test/test_amqp.py
@@ -1,3 +1,4 @@
+import json
 import pytest
 from amqp import *
 
@@ -105,3 +106,49 @@ def test_consume_nomsg(com):
     assert res
     assert msg[0] == None
     assert mtag[0] == 0
+
+def test_batch_publish(com):
+    Q = "batch"
+    msgs = [
+        {"routingKey":"#", "body":"message1"},
+        {"routingKey":"#", "body":"message2", "properties":{"AppId":"MYAPP","MessageId":"message2"}, "headers":{"some-header":13}},
+        {"routingKey":"#", "body":"message3", "properties":{"AppId":"MYAPP","MessageId":"message3"}, "headers":{"some-header":13}},
+    ]
+    bind_queue(com, Q)
+    res = com.call_proc("BatchPublish", Q, False, json.dumps(msgs))
+    assert res
+    ctag = consume(com, Q)
+    for i in range(3):
+        msg = [""]
+        mtag = [-1]
+        res, ret = com.call_func("BasicConsumeMessage", ctag, msg, mtag, 10000)
+        assert res
+        assert ret
+        assert msg[0] == "message" + str(i+1)
+        res, val = com.get_prop("MessageId")
+        assert res
+        assert i == 0 or val == msg[0]
+        res, val = com.get_prop("AppId")
+        assert res
+        assert i == 0 or val == "MYAPP"
+        if i > 0:
+            res, ret = com.call_func("GetHeaders")
+            assert res
+            hdr = json.loads(ret)
+            assert hdr['some-header'] == 13
+        com.call_proc("BasicAck", mtag[0])
+
+def test_queue_message_count(com):
+    Q = "count_queue"
+    make_queue(com, Q)
+    res, count = com.call_func("GetQueueMessageCount", Q)
+    assert res
+    assert count == 0
+    publish(com, Q, "one")
+    res, count = com.call_func("GetQueueMessageCount", Q)
+    assert res
+    assert count == 1
+    publish(com, Q, "two")
+    res, count = com.call_func("GetQueueMessageCount", Q)
+    assert res
+    assert count == 2